core/Makefile: Fix filenames with commas
Prior to this, any filenames with commas, for example device specific
nvrams or hcds compatible with the mainline kernel, would cause a build
failure because GNU make can't handle passing an argument with a comma
because comma is the delimiter.

To work around this pass the arguments via variables local to the rule.

Change-Id: I852b36d7194394389a63683fc8c68474f8323d9b
Signed-off-by: Jis G Jacob <studiokeys@blissroms.org>